requirements.txt
django==4.2.1
antlr4-python3-runtime==4.13.0
whitenoise>=6.0
unidecode